+++ Project/GameScreen.cpp
@@ -17,11 +17,11 @@
 #include <algorithm>
 #include <iostream>
 #include <memory>
-#include "BorderLine.h";
+#include "BorderLine.h"
 
 GameScreen::GameScreen(std::shared_ptr<Client> client)
 	: m_client(client) {
-	m_sm.addScreen(HUD_SCREEN, std::make_shared<HudScreen>(sf::Vector2f(WINDOW_SIZE_X, WINDOW_SIZE_Y), m_player));
+	m_sm.addScreen(HUD_SCREEN, std::make_shared<HudScreen>(sf::Vector2f((float)WINDOW_SIZE_X, (float)WINDOW_SIZE_Y), m_player));
 
 	m_sm.setScreen(HUD_SCREEN);
 
@@ -226,7 +226,7 @@ void GameScreen::update(ConnectionInfo & pi) {
 	m_view.setCenter(m_player->getCenter());
 	//Update server - start pos
 	Updates<std::shared_ptr<PlayerInfo>, Request>::getInstance().add(m_player->getPlayerInfo());
-	m_sm.addScreen(SHOP_SCREEN, std::make_shared<ShopScreen>(sf::Vector2f(WINDOW_SIZE_X, WINDOW_SIZE_Y), m_player));
+	m_sm.addScreen(SHOP_SCREEN, std::make_shared<ShopScreen>(sf::Vector2f((float)WINDOW_SIZE_X, (float)WINDOW_SIZE_Y), m_player));
 }
 
 void GameScreen::updateFromServer() {
